move paper image view into a separate right side function component
